feat(wms): 增加发货单反审核和删除的校验
- 新增校验逻辑，禁止删除或反审核已生成物流单或报关单的发货单
- 在 ApiError 枚举中添加相关错误信息
- 优化物流产品信息保存逻辑，自动设置海关状态为已完成

+++ erp-server/erp-server-wms/src/main/java/com/erp/server/wms/service/impl/FirstMileDeliveryServiceImpl.java
@@ -911,6 +911,16 @@ public class FirstMileDeliveryServiceImpl extends SuperServiceImpl<FirstMileDeli
         if (CollUtil.isNotEmpty(inboundEntityList)){
             throw new ServiceException("存在下游海外入库单【{}】，禁止删除", inboundEntityList.stream().map(OverseasWarehouseInboundEntity::getCode).filter(CharSequenceUtil::isNotBlank).collect(Collectors.joining(",")));
         }
+
+        //校验下游单据是否生成【包含报关单，物流单】状态为已生成 不可反审核【提示：报关单/物流单[单号]已生成，不可反审核】
+        List<LogisticsBillEntity> tmsFirstMileLogisticEntities = tmsFirstMileLogisticFeign.listByOutstockIds(Collections.singletonList(entity.getId()));
+        if (CollUtil.isNotEmpty(tmsFirstMileLogisticEntities)) {
+            throw new ServiceException(ApiError.TMS_FIRST_MILE_LOGISTIC_EXISTS_NOT_DEL, tmsFirstMileLogisticEntities.get(0).getTransportNo());
+        }
+        List<TmsDeclareBillEntity> tmsDeclareBillEntities = tmsDeclareBillFeign.listBySourceIds(Collections.singletonList(entity.getId()));
+        if (CollUtil.isNotEmpty(tmsDeclareBillEntities)) {
+            throw new ServiceException(ApiError.TMS_DECLARE_BILL_EXISTS_NOT_DEL, tmsDeclareBillEntities.get(0).getCode());
+        }
         //删除装箱信息
         if(Objects.nonNull(packingTask)){
             packingTaskService.delete(packingTask);
